BUG#50767: Some RPL tests started to fail in next-mr-merge on Linux x86_64 debug
Two test cases fail because the suppression for the unsafe
warning needs to be updated (BUG@39934 refactored this part and
these changes are only in mysql-next-mr - this is why we notice
them now when merging in next-mr). This is the case for
rpl_nondeterministic_functions and rpl_misc_functions test
cases. rpl_stm_binlog_direct test case is not needed in version >
5.1. The rpl_heartbeat_basic test case fails because patch for
BUG@50397 removed the CHANGE MASTER in the slave that would set
it's period to 1/10 of the master. This would cause the test
assertion to fail.

The fixes for the issues described above are:

 - rpl_misc_functions - updated suppression message
 - rpl_nondeterministic_functions - updated suppression message
 - rpl_stm_binlog_direct - removed the test case (it is not 
                           needed in versions > 5.1)
 - rpl_heartbeat_basic - deployed instruction: 
   CHANGE MASTER TO MASTER_HEARTBEAT_PERIOD=0.1;

--BZR--
revision-id: alik@sun.com-20100202092628-ddyr2iic91h910b1
property-branch-nick: mysql-next-mr-merge
testament3-sha1: 84ddd6a01259fb5d6b0e921bf9286e88e08bce4a