Bug #50907 Assertion `hash_tables->table->next == __null' on            HANDLER OPEN
The problem was a too restrictive assert in the code for 
HANDLER ... OPEN and HANDLER ... READ that checked table->next
to verify that we didn't open views or merge tables.

This pointer is also used to link temporary tables together
(see thd->temporary_tables). In this case TABLE::next can be
set even if we're trying to open a single table.

This patch adjust the two asserts to also check for the presence
of temporary tables.

Test case added to handler_myisam.test.

--BZR--
revision-id: jon.hauglid@sun.com-20100205145217-3b7n3jfx1y47z41u
property-branch-nick: mysql-next-4284-bug50907
testament3-sha1: 9be372f2322d0e0b148dba6caa26683989e12505