BUG#50984: check_testcase fails for rpl_tmp_table_and_DDL
We found that there are some tests that are not cleaning
up properly:
      
  1. rpl_tmp_table_and_DDL
  2. rpl_do_grant
  3. rpl_sync
      
For #1 and #2 we found that the slave would not, for some
cases, replicate all the instructions the master processed 
in the cleanup section. We fix these by deploying some 
synchronization commands in the test cases so that slave 
processes all clean up instructions.
      
As for #3, this is tracked as part of another bug 
(BUG@50442).
--BZR--
revision-id: luis.soares@sun.com-20100210121255-zx028qhq2jksqetv
property-branch-nick: mysql-trunk-bugfixing
testament3-sha1: fd1399b3ad8c0362004d98f5e2e07a7b7a1211bc