#51466 : Source packages are broken with cmake in a cmake-agnostic environment
In the worst case possible scenario (no bzr, in-source build),
make dist produced a package that compiled ok with autotools
but failed to package because extra make_binary_distribution was
found in source package and was not built. make_binary_distribution
contained paths of the build machine.

Fix:  exclude some scripts that are produced in cmake build. 
Note that there is no good general fix for it in this specific scenario.
it is advisable to build source packages out of source or in bzr repo.
--BZR--
revision-id: wlad@sun.com-20100224190705-alv4xe1kfetq7rh4
property-branch-nick: mysql-next-mr-bugfixing
testament3-sha1: 6924b2afba3f466babc871360995f0334fafca7b