Bug#49417 some complaints about mysqld --help --verbose output
This patch fixes some typos and poorly formulated sentences in
the output from mysqld --help --verbose.

Some of the problems described in the bug report are already
handled by the patch for Bug#49447, and are therefore not
included in this patch.
--BZR--
revision-id: staale.smedseng@sun.com-20100301134951-1so7n5vi7xu0brf0
property-branch-nick: 49417-51
testament3-sha1: 930a66f59322e89b18f6ed962d7987bc1f104253