Bug#51612 Un initialized locker_lost variable in pfs_instr.cc
Fixed the missing initialization of locker_lost.

This fix is not strictly necessary, but is desirable to re-align the code
from 5.5 and 6.0, and reduce the spurious code differences.

This will facilitate maintenance and help to apply patches cleanly, for merges.

--BZR--
revision-id: marc.alff@sun.com-20100302001001-73lojr3g8l0hxva3
property-branch-nick: mysql-next-mr-bugfixing
testament3-sha1: 81cc29f576f9bd6640dae49a1fc242699fd25b16