branches/5.1:
In innodb_bug38231.test: replace the fragile sleep 0.2 that depends on timing
with a more robust condition which waits for the TRUNCATE and LOCK commands
to appear in information_schema.processlist. This could also break if there
are other sessions executing the same SQL commands, but there are none during
the execution of the mysql test.

--BZR--
revision-id: svn-v4:cee13dc7-1704-0410-992b-c9b4543f1246:branches/5.1:6788
testament3-sha1: 149794d3bc8d4209b34d9358aea61a097b1e6a30