Bug#50918: Date columns treated differently in Views than in Base Tables
The type inferrence of a view column caused the result to be
interpreted as the wrong type: DATE colums were interpreted
as TIME and TIME as DATETIME. This happened because view
columns are represented by Item_ref objects as opposed to
Item_field's. Item_ref had no method for retrieving a TIME
value and thus was forced to depend on the default
implementation for any expression, which caused the
expression to be evaluated as a string and then parsed into
a TIME/DATETIME value.

Fixed by letting Item_ref classes forward the request for a
TIME value to the referred Item - which is a field in this
case - this reads the TIME value directly without
conversion.

--BZR--
revision-id: martin.hansson@sun.com-20100316092007-1s5ydd7ho0ugakhv
property-branch-nick: 5.1bt-commit
testament3-sha1: 6b1b39788082942b435043969ab15343cba02108