BUG#52202: mysqlbinlog_row* fail in daily-trunk on Sol10 x86_64            debug_max
There was a buffer overrun when unpacking the date
field. Incidentaly, this seems to affect only solaris x86_64
debug builds, but others platforms may be vulnerable as well.

In particular, the buffer size used was not taking into
consideration that the '\0' character would be written into
it.

We fix this by increasing the size of the buffer used to
accommodate one extra byte (the one for the '\0').
--BZR--
revision-id: luis.soares@sun.com-20100325003016-m4p8s0d3l1m66s6p
property-branch-nick: mysql-trunk-bugfixing
testament3-sha1: b7ba032f5e009e21cc93fa887c9c42a0f292400b