Applying InnoDB snapshot
Detailed revision comments:

r6740 | sunny | 2010-02-18 13:44:31 +0200 (Thu, 18 Feb 2010) | 6 lines
branches/zip: Don't print the entire lock bit set if the block was not
found in the buffer pool. Only print the bits that are set and that
information is in the lock and not in the block.

See rb://256 approved by Marko.

--BZR--
revision-id: svoj@sun.com-20100401115848-blftc2x3z63qov4k
property-branch-nick: mysql-5.1-bugteam
testament3-sha1: 03266317d214513f21b5cf972c1e038aac015000