Move the InnoDB Plugin tests from storage/innodb_plugin/mysql-test/ where they are ignored to a new test suite "innodb_plugin".
Remove a hack in mtr that was deployed to run the builtin InnoDB tests against
the InnoDB Plugin. Also detect if a test is an 'innodb plugin test' and if so
then transparently replace the builtin InnoDB with the InnoDB Plugin.

--BZR--
revision-id: vasil.dimov@oracle.com-20100409125651-6hqcjfwlh9qi2e82
property-branch-nick: mysql-5.1-innodb
testament3-sha1: 6ab71c78be1725de33c46d800fbaaa9771202b2c