A patch for Bug#52444 (mysql_upgrade fails b/w 5.1 -> 5.5 (Celosia)).
The problem was that the code that works with stored routines was present
in mysql_system_tables.sql. This is wrong because until upgrade is finished,
stored routines may be (and were) not available.

The fix is to move that code to the end of mysql_system_tables_fix.sql.
--BZR--
revision-id: alik@sun.com-20100414164004-uyo6q0b9ftqqx3se
property-branch-nick: mysql-trunk-bf-bug52444
testament3-sha1: 188d4c899116484729d6fbb90239a1286f9c912a