Bug #51327 MyISAM table is automatically repaired on ALTER            even if myisam-recover is OFF
The problem was that a corrupted MyISAM table was auto repaired
even if the myisam_recover_options server variable (or the 
myisam_recover option) was set to OFF.

The reason was that the auto_repair() function, which is supposed
to say if auto repair is to be used, did not use the server variable
setting correctly. This bug was a regression introduced by WL#4738.

This patch fixes the problem by making sure auto_repair() returns
FALSE if myisam_recover_options is set to OFF.

Test case added to myisam.test.

--BZR--
revision-id: jon.hauglid@sun.com-20100415071629-ntpie9jvg1nuw3bi
property-branch-nick: mysql-trunk-runtime-bug51327
testament3-sha1: 18fe9bdd41b0890b1fb143071826322981acfc2a