BUG#53101: binlog.binlog_tmp_table fails on sol10 sparc64 max in            parallel mode
The failure has nothing to do with parallel, but rather on the
order the tests are executed. In this case, the test
binlog_tmp_table (lets call it test2) was not ensuring that the
binary logs would be reset when it started. Later the test issues
a mysqlbinlog .../master-bin.000002 | mysql ... If the test that
was executed before this one (lets call it test1) had issued a
flush logs, then the file in use in test1 (master-bin.000002)
would not actually match the one that was expected. Eventually,
this would cause the statements logged in test1 to be replayed,
instead of the ones logged in the beginning of test2.
    
We fix this by:
  1. adding RESET MASTER to the beginning of binlog_tmp_table
  2. setting dynamically the file to use in binlog_tmp_table
     
Only #1 was needed, but the two make the tests cases more robust.
--BZR--
revision-id: luis.soares@sun.com-20100427112555-kpa00w8phl4arv91
property-branch-nick: mysql-trunk-bugfixing-push
testament3-sha1: 5f21d575b8bd406626db1bbb4c7062d232c7971e