Bug #47453: InnoDB incorrectly changes TIMESTAMP columns when JOINed during an UPDATE
Extended the fix for bug 29310 to multi-table update:

When a table is being updated it has two set of fields - fields required for
checks of conditions and fields to be updated. A storage engine is allowed
not to retrieve columns marked for update. Due to this fact records can't
be compared to see whether the data has been changed or not. This makes the
server always update records independently of data change.
  
Now when an auto-updatable timestamp field is present and server sees that
a table handle isn't going to retrieve write-only fields then all of such
fields are marked as to be read to force the handler to retrieve them.
--BZR--
revision-id: joro@sun.com-20100428125554-sympdx16ghfmlh7a
property-branch-nick: B47453-5.1-bugteam
property-rebase-of: joro@sun.com-20100428125554-fexcobfoo4vcaak3
testament3-sha1: 8822b0ede37f5933d82125fc827978ed02300108