Bug#48800 CREATE TABLE t...SELECT fails if t is a temporary           table
If a temporary table A exists, and a (permanent) table 
with the same name is attempted created with 
"CREATE TABLE ... AS SELECT", the create would fail with 
an error.
   1050: Table 'A' already exists

The error occured in MySQL 5.1 releases, but is not
present in MySQL 5.5.   This patch adds a regression
test to ensure that the problem does not reoccur.

--BZR--
revision-id: magne.mahre@sun.com-20100505101707-hbuahi9sq2zgobai
property-branch-nick: mysql-trunk-runtime
testament3-sha1: b6a6c1e7511e72cedc03a3838a15e4b3b15fc035