Bug#52357: Assertion failed: join->best_read in greedy_search optimizer_search_depth=0
The algorithm inside restore_prev_nj_state failed to
properly update the counters within the NESTED_JOIN
tree. The counter was decremented each time a table in the
node was removed from the QEP, the correct thing to do being
only to decrement it when the last table in the child node
was removed from the plan. This lead to node counters
getting negative values and the plan thus appeared
impossible. An assertion caught this.

Fixed by not recursing up the tree unless the last table in
the join nest node is removed from the plan
--BZR--
revision-id: martin.hansson@sun.com-20100506084500-8evafe6ghrib6zrm
property-branch-nick: 5.0bt
testament3-sha1: 186e9817d44c1cd131bd371f9cc9a91302d13183