Bug #53798 OPTIMIZE TABLE breaks repeatable read
The problem was that OPTMIZE TABLE was allowed to run on a table
in use by a transaction in a different connection. This caused
repeatable read to break.

This bug was fixed by the introduction of metadata locking, WL#4284.
OPTIMIZE TABLE will now be blocked until the transaction using the
table, has ended.

This patch contains a regression test added to innodb_mysql_lock.test
and no code changes.

--BZR--
revision-id: jon.hauglid@sun.com-20100519113221-47d2ye4me5rgbj0q
property-branch-nick: mysql-trunk-runtime-newbug
testament3-sha1: dee240f96dffaf237f9a02fa3630843721960cd6